Update the GCN model used
