+++ yeaz/nns/weights/results_fourier_10_and_f10_locality_False/_100KB_/2023-12-19_12:47:30/hyperparams.json
@@ -1,0 +1,1 @@
+{"num_node_attr": 60, "num_edge_attr": 10, "dropout_rate": 0.01, "encoder_hidden_channels": 120, "encoder_num_layers": 4, "conv_hidden_channels": 120, "conv_num_layers": 3, "num_classes": 1, "lr": 0.001, "weight_decay": 0.01}
