+++ yeaz/nns/weights/results_fourier_10_and_f10_locality_False/_100KB_/2023-12-19_12:47:30/metadata.json
@@ -1,0 +1,1 @@
+{"cell_graph_config": {"cell_f_list": ["fourier", 10, false, ["area", "r_equiv", "r_maj", "r_min", "angel", "ecc", "maj_x", "maj_y", "min_x", "min_y"]], "edge_f_list": ["cmtocm_x", "cmtocm_y", "cmtocm_len", "cmtocm_angle", "contour_dist"], "scale_time": 1, "nn_threshold": 12, "output_folder": "/mlodata1/hokarami/fari/tracking/generated_data/scaled_images/cell_graphs_fourier_10_f10_locality_False", "input_segmentations": ["/mlodata1/hokarami/fari/bread/data/segmentations/colony000_segmentation.h5", "/mlodata1/hokarami/fari/bread/data/segmentations/colony006_segmentation.h5", "/mlodata1/hokarami/fari/bread/data/segmentations/colony008_segmentation.h5", "/mlodata1/hokarami/fari/bread/data/segmentations/colony007_segmentation.h5"]}, "ass_graph_config": {"cellgraph_dirs": ["/mlodata1/hokarami/fari/bread/data/generated/test_trcking/cell_graphs_fourier_10_and_f10_locality_False/colony007_segmentation/", "/mlodata1/hokarami/fari/tracking/generated_data/scaled_images/cell_graphs_fourier_10_f10_locality_False/colony008_segmentation/"], "output_folder": "/mlodata1/hokarami/fari/tracking/generated_data/scaled_images/ass_graphs_fourier_10_f10_locality_False", "framediff_min": 1, "framediff_max": 4, "t1_max": 120, "t1_min": 0}, "train_config": {"ass_graphs": "/mlodata1/hokarami/fari/tracking/generated_data/scaled_images/ass_graphs_fourier_10_f10_locality_False", "result_dir": "/mlodata1/hokarami/fari/tracking/results/scaled_images/results_fourier_10_and_f10_locality_False", "dataset": "colony_0123478_test_set_1234567_dt_1234_t_all", "valid_dataset": "colony001", "min_file_kb": 100, "filter_file_mb": 30, "dropout_rate": 0.01, "encoder_hidden_channels": 120, "encoder_num_layers": 4, "conv_hidden_channels": 120, "conv_num_layers": 3, "max_epochs": 50, "lr": 0.001, "weight_decay": 0.01, "step_size": 1024, "gamma": 0.5, "cv": 1, "patience": 6, "scoring": "valid_f1_ass"}, "test_config": {"ass_graphs": "/mlodata1/hokarami/fari/tracking/generated_data/scaled_images/ass_graphs_fourier_10_f10_locality_False", "result_dir": "/mlodata1/hokarami/fari/tracking/results/scaled_images/results_fourier_10_and_f10_locality_False", "dataset": "colony_56__dt_1234__t_all", "filter_file_mb": 30}, "use_wandb": true}
