+++ yeaz/nns/weights/results_fourier_10_and_f10_locality_False/_100KB_/2023-12-19_12:47:30/result_sum_colony_56__dt_1234__t_all_filtered.csv
@@ -1,0 +1,9 @@
+timediff,colony,f1,tp,fp,fn,tn,num_cells1,num_cells2,precision,recall
+5,5,1.0,2684,0,0,57472,2684,2720,1.0,1.0
+5,6,0.9997030878859857,3367,1,1,229059,3368,3496,0.9997030878859857,0.9997030878859857
+10,5,0.9996220710506425,2645,1,1,56756,2646,2718,0.9996220710506425,0.9996220710506425
+10,6,0.9981487195310089,3235,6,6,220606,3241,3493,0.9981487195310089,0.9981487195310089
+15,5,0.9976993865030674,2602,6,6,56030,2608,2716,0.9976993865030674,0.9976993865030674
+15,6,0.9913406029506093,3091,27,27,212356,3118,3490,0.9913406029506093,0.9913406029506093
+20,5,0.9945546479968884,2557,14,14,55330,2571,2713,0.9945546479968884,0.9945546479968884
+20,6,0.9776517678452301,2931,67,67,204187,2998,3487,0.9776517678452301,0.9776517678452301
